backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
chromadb==0.4.18
pillow==10.1.0
python-dotenv==1.0.0
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0
numpy==1.24.3
